requirements.txt
numpy==1.25.0
pandas==2.1.2    
psycopg2==2.9.7    # สำหรับเชื่อมต่อ PostgreSQL
boto3==1.28.0   # สำหรับเชื่อมต่อ MinIO
apache-airflow-providers-amazon==8.6.0
